backend/requirements_full_deployment.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
asgiref==3.6.0
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
bangla==0.0.2
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
codegen==1.0
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dataclasses-json==0.6.6
dateparser==1.1.8
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
Django==4.2
django-cors-headers==4.3.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
docopt==0.6.2
docutils==0.20.1
einops==0.8.0
encodec==0.1.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
g2pkk==0.1.2
greenlet==3.0.3
grpcio==1.64.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
gunicorn==22.0.0
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
imagesize==1.4.1
inflect==7.2.1
iniconfig==2.0.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llama-index-core==0.10.33
llama-index-embeddings-huggingface==0.2.0
llama-index-llms-ollama==0.1.3
llama-index-readers-file==0.1.19
llamaindex-py-client==0.1.19
llvmlite==0.42.0
marisa-trie==1.1.1
Markdown==3.6
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
minijinja==2.0.1
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
num2words==0.5.13
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openai==1.30.5
packaging==23.1
pandas==1.5.3
patterns==0.3
Pillow==9.5.0
platformdirs==4.2.0
pluggy==1.0.0
pooch==1.8.1
preshed==3.0.9
protobuf==5.27.0
psutil==5.9.8
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.17.2
pynndescent==0.5.12
pyparsing==3.1.2
pypdf==4.2.0
pypinyin==0.51.0
pysbd==0.3.4
pytest==7.3.1
pytest-django==4.5.2
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==2.7.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
snowballstemmer==2.2.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.30
sqlparse==0.4.4
srsly==2.4.8
striprtf==0.0.26
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy==1.12.1
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
tqdm==4.66.4
trainer==0.0.36
transformers==4.41.2
triton==2.3.0
TTS==0.22.0
typeguard==4.3.0
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2023.3
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.1
virtualenv==20.25.1
wasabi==1.1.3
weasel==0.3.4
Werkzeug==3.0.3
whitenoise==6.5.0
wrapt==1.16.0
yarl==1.9.4
